requirements.txt
strawberry-graphql
uvicorn
fastapi
sqlalchemy
mysqlclient
pyjwt
bcrypt
python-dotenv
